feat(ios): remove not needed available and split up code;
