+++ Sources/App/Settings/Eureka/VoiceShortcutRow.swift
@@ -3,7 +3,6 @@ import Foundation
 import IntentsUI
 import Shared
 
-@available(iOS 13.0, *)
 public class VoiceShortcutCell: Cell<INShortcut>, CellType,
     INUIAddVoiceShortcutButtonDelegate,
     INUIAddVoiceShortcutViewControllerDelegate,
@@ -90,7 +89,6 @@ public class VoiceShortcutCell: Cell<INShortcut>, CellType,
     }
 }
 
-@available(iOS 13.0, *)
 public final class VoiceShortcutRow: Row<VoiceShortcutCell>, RowType {
     var buttonStyle: INUIAddVoiceShortcutButtonStyle = .automatic
