+++ Sources/Shared/API/Responses/ConfigResponse.swift
