+++ Sources/Shared/API/Responses/MobileAppConfig/MobileAppConfig.swift
@@ -1,0 +1,17 @@
+import Foundation
+import ObjectMapper
+
+public struct MobileAppConfig: ImmutableMappable {
+    public var push: MobileAppConfigPush
+    public var actions: [MobileAppConfigAction]
+
+    internal init(push: MobileAppConfigPush = .init(), actions: [MobileAppConfigAction] = []) {
+        self.push = push
+        self.actions = actions
+    }
+
+    public init(map: Map) throws {
+        self.push = (try? map.value("push")) ?? MobileAppConfigPush()
+        self.actions = map.value("actions", default: [])
+    }
+}
