+++ Sources/Shared/API/Responses/MobileAppConfig/MobileAppConfigPush.swift
@@ -1,0 +1,14 @@
+import Foundation
+import ObjectMapper
+
+public struct MobileAppConfigPush: ImmutableMappable {
+    public var categories: [MobileAppConfigPushCategory]
+
+    internal init(categories: [MobileAppConfigPushCategory] = []) {
+        self.categories = []
+    }
+
+    public init(map: Map) throws {
+        self.categories = map.value("categories", default: [])
+    }
+}
