+++ Sources/Shared/API/Responses/MobileAppConfig/MobileAppConfigPushCategory.swift
@@ -1,30 +1,6 @@
 import Foundation
 import ObjectMapper
 
-public struct MobileAppConfigAction: ImmutableMappable, UpdatableModelSource {
-    var name: String
-    var backgroundColor: String?
-    var labelText: String?
-    var labelColor: String?
-    var iconIcon: String?
-    var iconColor: String?
-    var showInCarPlay: Bool?
-    var showInWatch: Bool?
-
-    public init(map: Map) throws {
-        self.name = try map.value("name")
-        self.backgroundColor = try? map.value("background_color")
-        self.labelText = try? map.value("label.text")
-        self.labelColor = try? map.value("label.color")
-        self.iconIcon = try? map.value("icon.icon")
-        self.iconColor = try? map.value("icon.color")
-        self.showInCarPlay = try? map.value("show_in_carplay")
-        self.showInWatch = try? map.value("show_in_watch")
-    }
-
-    public var primaryKey: String { name }
-}
-
 public struct MobileAppConfigPushCategory: ImmutableMappable, UpdatableModelSource {
     public struct Action: ImmutableMappable {
         public var title: String
@@ -76,31 +52,4 @@ public struct MobileAppConfigPushCategory: ImmutableMappable, UpdatableModelSour
     }
 
     public var primaryKey: String { identifier.uppercased() }
-}
-
-public struct MobileAppConfigPush: ImmutableMappable {
-    public var categories: [MobileAppConfigPushCategory]
-
-    init(categories: [MobileAppConfigPushCategory] = []) {
-        self.categories = []
-    }
-
-    public init(map: Map) throws {
-        self.categories = map.value("categories", default: [])
-    }
-}
-
-public struct MobileAppConfig: ImmutableMappable {
-    public var push: MobileAppConfigPush
-    public var actions: [MobileAppConfigAction]
-
-    init(push: MobileAppConfigPush = .init(), actions: [MobileAppConfigAction] = []) {
-        self.push = push
-        self.actions = actions
-    }
-
-    public init(map: Map) throws {
-        self.push = (try? map.value("push")) ?? MobileAppConfigPush()
-        self.actions = map.value("actions", default: [])
-    }
 }
